rename the api/embedding route, since it is now model-independent
